@@ -1,12 +1,12 @@
 plugins {
-	id 'org.springframework.boot' version '2.4.1'
+	id 'org.springframework.boot' version '3.1.7'
 	id 'io.spring.dependency-management' version '1.0.10.RELEASE'
 	id 'java'
 	id "maven-publish"
 }
 group = 'jpabook'
 version = '0.0.1-SNAPSHOT'
-sourceCompatibility = '11'
+sourceCompatibility = '17'
 configurations {
 	compileOnly {
 		extendsFrom annotationProcessor
