@@ -1,7 +1,7 @@
 package jpabook.jpashop.domain;
 
 
-import javax.persistence.*;
+import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.Setter;
 
